backend/requirements.txt
setuptools==58.1.0
wheel==0.41.3
paho-mqtt==2.1.0
opencv-python==4.10.0.84
Pillow==11.3.0
numpy==1.26.0
redis==4.2.2
twilio==9.6.5
torch==2.7.1
rembg==2.0.67
scikit-learn==1.7.1
onnxruntime==1.22.1
mysql-connector-python==9.4.0
mediapipe==0.10.21
insightface[full]
fastapi==0.95.2
dotenv==0.9.9
uvicorn==0.35.0
ultralytics==8.3.107